Split build-logic module into commons and regtest
