--- build-logic/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id 'groovy-gradle-plugin'
-    id 'org.gradle.kotlin.kotlin-dsl' version '2.1.7'
-}
-
-repositories {
-    mavenCentral()
-}
-
-gradlePlugin {
-    plugins {
-        simplePlugin {
-            id = 'bisq.gradle.BisqPlugin'
-            implementationClass = 'bisq.gradle.BisqPlugin'
-        }
-    }
-}
