+++ build-logic/commons/src/main/groovy/bisq.protobuf.gradle
