+++ build-logic/regtest/build.gradle
@@ -1,0 +1,16 @@
+plugins {
+    id 'org.gradle.kotlin.kotlin-dsl' version '2.1.7'
+}
+
+repositories {
+    mavenCentral()
+}
+
+gradlePlugin {
+    plugins {
+        simplePlugin {
+            id = 'bisq.gradle.regtest.BisqRegtestPlugin'
+            implementationClass = 'bisq.gradle.regtest.BisqRegtestPlugin'
+        }
+    }
+}
