+++ build-logic/regtest/src/main/kotlin/bisq/gradle/regtest/elementsd/ElementsdRpcClient.kt
@@ -1,0 +1,26 @@
+package bisq.gradle.regtest.elementsd
+
+object ElementsdRpcClient {
+    private val cmdArgs = listOf(
+        "elements-cli",
+        "-rpcwait",
+        "-chain=elementsregtest",
+
+        "-rpcuser=bisq",
+        "-rpcpassword=bisq"
+    )
+
+    fun daemonRpcCall(args: List<String>): String {
+        val process = ProcessBuilder(cmdArgs + args)
+            .start()
+        process.waitFor()
+
+        return process.inputStream
+            .bufferedReader()
+            .readText()
+    }
+
+    fun walletRpcCall(walletPath: String, args: List<String>): String {
+        return daemonRpcCall(listOf("-rpcwallet=$walletPath") + args)
+    }
+}
