+++ build-logic/settings.gradle
@@ -1,0 +1,2 @@
+include 'commons'
+include 'regtest'
